Finalisation of menu, finalisation of ranger animations, start screen addition
The menu is now completely functional, with help and high-score screens. Switches to map scene when the "play game" option is selected.
Added a start screen, stating that the game does not use a mouse to ease confusion. This screen switches to the menu scene after some time.
Ranger animations have been finalised and coded.